fixed the login and signup page and unscrollable
